cleanly refactor methods
