+++ package.json
@@ -4,6 +4,7 @@
   "description": "Generate autocomplete for Salesforce Aura code snippet",
   "scripts": {
     "test": "./node_modules/mocha/bin/mocha test",
+    "clean": "rm -f snippet/aura*",
     "pkg": "node generateJsAutoComplete.js --silent; node generateXmlAutoComplete.js --silent",
     "start": "node generateJsAutoComplete.js; node generateXmlAutoComplete.js"
   },
