+++ snippet/aura.attributes.atom.cson
@@ -1,0 +1,1 @@
+undefined
