+++ snippet/aura.event.js.atom.cson
@@ -1,0 +1,1 @@
+undefined
