+++ snippet/aura.event.js.sublime-completions
