+++ snippet/aura.js.atom.cson
