+++ snippet/aura.js.sublime-completions
