+++ snippet/aura.uitags.atom.cson
@@ -1,0 +1,1 @@
+undefined
